feat: add client update

+++ projeto 2/code/src/main/java/com/example/project_2/controller/UserController.java
@@ -6,17 +6,15 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.web.bind.annotation.*;
 
-import com.example.project_2.model.Vehicle;
 import com.example.project_2.model.DTO.UserDTO;
 import com.example.project_2.model.DTO.VehicleDTO;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import java.util.List;
-import com.example.project_2.Enums.VehicleStatus;
+import com.example.project_2.Util.Auth;
 
 @RestController
 @RequestMapping("/api/users")
@@ -29,6 +27,9 @@ public class UserController {
 
     @Autowired
     private UserDTO userDTO;
+
+    @Autowired
+    private Auth auth;
 
     @Operation(summary = "Listar todos os usuários", description = "Retorna uma lista de todos os usuários cadastrados")
     @ApiResponses(value = {
@@ -48,6 +49,10 @@ public class UserController {
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteUser(@PathVariable Long id,
             @RequestHeader String token) {
+
+        if (!auth.authenticate(token)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Erro ao autenticar usuário");
+        }
 
         boolean deleted = userDTO.deleteUser(id, token);
 
@@ -79,27 +84,6 @@ public class UserController {
         }
     }
 
-    @Operation(summary = "Alterar senha do usuário", description = "Altera a senha de um usuário existente")
-    @ApiResponses(value = { @ApiResponse(responseCode = "200", description = "Senha alterada comsucesso"),
-            @ApiResponse(responseCode = "400", description = "Falha na alteração dasenha"),
-            @ApiResponse(responseCode = "404", description = "Usuário não encontrado"),
-            @ApiResponse(responseCode = "403", description = "Erro ao autenticarusuário")
-    })
-    @PutMapping("/change-password")
-    public ResponseEntity<String> changePassword(
-            @RequestBody PasswordChangeRequest request,
-            @RequestHeader String token) {
-
-        boolean changed = userDTO.changePassword(request.getNewPassword(), token);
-
-        if (changed) {
-            return ResponseEntity.ok("Senha alterada com sucesso");
-        } else {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body("Usuário não encontrado ou senha não pôde ser alterada");
-        }
-    }
-
     @Operation(summary = "Adicionar veículo ao usuário", description = "Adiciona um veículo ao usuário")
     @ApiResponses(value = { @ApiResponse(responseCode = "200", description = "Veículo adicionado com sucesso"),
             @ApiResponse(responseCode = "403", description = "Erro ao autenticar usuário")
@@ -123,19 +107,6 @@ public class UserController {
             e.printStackTrace();
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                     .body("Erro ao adicionar veículo: " + e.getMessage());
-        }
-    }
-
-    private static class PasswordChangeRequest {
-        private String oldPassword;
-        private String newPassword;
-
-        public String getOldPassword() {
-            return oldPassword;
-        }
-
-        public String getNewPassword() {
-            return newPassword;
         }
     }
